Add auto white balance and fix not able to write white balance
Fix the problem of not being able to set white balance using Property.
When trying to set white balance on my FL3-U3-13E4C-C, both this ros
driver and flycap software cannot set the white balance naturally.
After playing around with the flycap software, I found that I have
to set the highest bit of register 80C (which is white balance) to 1
to enable this feature. So I modified the original SetWhiteBalance to
use WriteRegister directly. And add support for auto white balance.